third_party/blink/web_tests/virtual/scoped-custom-element-registry-disabled/external/wpt/custom-elements/registries/ShadowRoot-init-customElementRegistry-expected.txt

This is a testharness.js-based test.
[FAIL] A newly attached disconnected ShadowRoot should use the global registry by default
  assert_equals: expected (object) object "[object CustomElementRegistry]" but got (undefined) undefined
[FAIL] A newly attached connected ShadowRoot should use the global registry by default
  assert_equals: expected (object) object "[object CustomElementRegistry]" but got (undefined) undefined
[FAIL] A newly attached disconnected ShadowRoot should use the scoped registry if explicitly specified in attachShadow
  Failed to construct 'CustomElementRegistry': Illegal constructor
[FAIL] A newly attached connected ShadowRoot should use the scoped registry if explicitly specified in attachShadow
  Failed to construct 'CustomElementRegistry': Illegal constructor
[FAIL] attachShadow() should use null registry when customElementRegistry is null (host uses global registry)
  assert_equals: expected (object) null but got (undefined) undefined
[FAIL] attachShadow() should use null registry when customElementRegistry is null (host uses custom registry)
  Failed to construct 'CustomElementRegistry': Illegal constructor
[FAIL] attchShadow on a builtin element with null customElementRegistry should create a ShadowRoot with null registry
  assert_equals: expected (object) null but got (undefined) undefined
[FAIL] attchShadow on a custom elememnt candidate with null customElementRegistry should create a ShadowRoot with null registry
  assert_equals: expected (object) null but got (undefined) undefined
[FAIL] attchShadow on a custom elememnt with null customElementRegistry should create a ShadowRoot with null registry
  assert_equals: expected (object) null but got (undefined) undefined
[FAIL] attachShadow() should use the null registry when the shadow host uses null registry and customElementRegistry is null
  Failed to construct 'CustomElementRegistry': Illegal constructor
Harness: the test ran to completion.
